IMessageListener, Connection Test Stubs, Mockito
Extracted IMessageListener from Connection, which will block whilst
listening for messages from a particular client.
Connection registers a callback to be triggered when a message is
received.
This allows me to test with in a single threaded environment, and gives
me a hook into the validateMessage which drives the the rest of the
class.
Messages can now be mocked in tests through a custom IMessageListener
implementation.
Started creating Connection test stubs.
Added Mockito to Maven dependencies.