Integration Tests Waiting Bug
Fixed a bug which meant the logic for waiting for server responses was
not working correctly.
This was due to the response callback not taking a reference to the test
but instead copying a reference to the count down latch. This meant that
when the tests created a new latch, the callback would still reference
the old latch. The callback now triggers a method on the test instance.